+++ src/main/java/dev/willyelton/crystal_tools/model/ModelSetup.java
@@ -1,7 +1,9 @@
 package dev.willyelton.crystal_tools.model;
 
 
+import dev.willyelton.crystal_tools.CrystalTools;
 import dev.willyelton.crystal_tools.Registration;
+import dev.willyelton.crystal_tools.levelable.tool.BowLevelableItem;
 import net.minecraft.client.model.EntityModel;
 import net.minecraft.client.model.geom.EntityModelSet;
 import net.minecraft.client.renderer.entity.ArmorStandRenderer;
@@ -18,21 +20,23 @@ import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 
-@Mod.EventBusSubscriber(modid = "crystal_tools", value = Dist.CLIENT, bus = Mod.EventBusSubscriber.Bus.MOD)
+@Mod.EventBusSubscriber(modid = CrystalTools.MODID, value = Dist.CLIENT, bus = Mod.EventBusSubscriber.Bus.MOD)
 public class ModelSetup {
     @SubscribeEvent
     public static void setModelProperties(FMLClientSetupEvent event) {
         // Bow things
-        ItemProperties.register(Registration.CRYSTAL_BOW.get(), new ResourceLocation("pull"), (p_174635_, p_174636_, p_174637_, p_174638_) -> {
-            if (p_174637_ == null) {
+        ItemProperties.register(Registration.CRYSTAL_BOW.get(), new ResourceLocation("pull"), (stack, level, entity, seed) -> {
+            if (entity == null) {
                 return 0.0F;
             } else {
-                return p_174637_.getUseItem() != p_174635_ ? 0.0F : (float)(p_174635_.getUseDuration() - p_174637_.getUseItemRemainingTicks()) / 20.0F;
+                if (stack.getItem() instanceof BowLevelableItem bowItem) {
+                    return entity.getUseItem() != stack ? 0.0F : (float)(stack.getUseDuration() - entity.getUseItemRemainingTicks()) / bowItem.getChargeTime(stack);
+                }
+                return 0;
             }
         });
-        ItemProperties.register(Registration.CRYSTAL_BOW.get(), new ResourceLocation("pulling"), (p_174630_, p_174631_, p_174632_, p_174633_) -> {
-            return p_174632_ != null && p_174632_.isUsingItem() && p_174632_.getUseItem() == p_174630_ ? 1.0F : 0.0F;
-        });
+        ItemProperties.register(Registration.CRYSTAL_BOW.get(), new ResourceLocation("pulling"),
+                (stack, level, entity, seed) -> entity != null && entity.isUsingItem() && entity.getUseItem() == stack ? 1.0F : 0.0F);
     }
 
     @SubscribeEvent